refac make some utils public
